beanstalk/requirements.txt
click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
Werkzeug==0.16.0
DateTime==4.3